+++ migrations/2.0_deploy_chain_storage.js
@@ -1,0 +1,15 @@
+const getConfig = require('./helpers/getConfig');
+
+const ChainStorageArtifact = artifacts.require('ChainStorage');
+
+module.exports = (deployer, network, accounts) => {
+  const { options, config } = getConfig(network, accounts);
+
+  return deployer.deploy(
+    ChainStorageArtifact,
+    config.Chain.blocksPerPhase,
+    config.Chain.minimumStakingTokenPercentage,
+    false,
+    options,
+  );
+};