+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+
+### Changed
+
+## [0.4.0] - 2019-03-14
+### Added
+- option to update `minimumStakingTokenPercentage`.
+- `initialBlockHeights` to set the initial block-height per shard
+- fetch VerifierRegistry contract from ContractRegistry
+
+### Changed
+- smart contracts are complient with solidity 0.5
+- unit tests are complient with web3 1.0
+- when verifier proposing, `Chain` checks, if verifier is active instead
+  of if it is only created.
+- split Chain contract into two separate contracts (storage and manager)
+
 ## [0.3.0] - 2018-12-13
 ### Added:
 - election verification
@@ -19,7 +37,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Chain interface
 - Getters for reading `blocks` structure
 - update of contract code to be consistent with Solidity 0.4.24
-- Remove commissioner and introduce election cycle with two phases: propose and reveal 
+- Remove commissioner and introduce election cycle with two phases: propose and reveal
 
 ## [0.1.0] - 2018-04-12
 ### Added: